should be consistent

+++ include/ctfa/include/ctfa/container/set.hpp
@@ -111,7 +111,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (!(obj < *it)) {
 				first = ++it;
 				count -= step + 1;
@@ -129,7 +129,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (!(obj < *it)) {
 				first = ++it;
 				count -= step + 1;
@@ -165,7 +165,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (*it < obj) {
 				first = ++it;
 				count -= step + 1;
