+++ single-header/ctre-unicode.hpp
@@ -4653,7 +4653,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (!(obj < *it)) {
 				first = ++it;
 				count -= step + 1;
@@ -4671,7 +4671,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (!(obj < *it)) {
 				first = ++it;
 				count -= step + 1;
@@ -4707,7 +4707,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (*it < obj) {
 				first = ++it;
 				count -= step + 1;
